add offscreen cells
Adds a configurable number of cells off the screen on each side.
This prevents incorrect behaviour due to fewer neighbours at the borders.

The more cells there are, the more 'correct' the game behaves.

However, due to only the direct neighbours being looked at for each cell a small number is sufficient,
as it's unlikely (though, still possible) that two spaceships collide off-screen and do a boomerang-like move
back to the visible game area